Fix vendoring hack
Fix the vendoring hack to override `sys.modules["archspec"]` only
if it isn't loaded already.  This can happen when the x86-64 plugin
is loaded as well, as it uses the same vendoring hack.  Once the module
is overwritten locally, Python is not adding the submodules to this new
module -- since they were added to the previous one already.

Signed-off-by: Michał Górny <mgorny@gentoo.org>